# Phase 2-20260109T205517Z-3-001/Phase 2/Phase 2/swagger-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComStag API",
    "description": "Organizations & Consumers platform",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://comstag-back.onrender.com",
      "description": "Generated server url"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/v1/testimonial": {
      "put": {
        "tags": ["update-testimonial-processor"],
        "summary": "Update testimonial",
        "description": "1. Update specific testimonial by id.\n2. Only consumer have access to this endpoint.\n3. Request code first.\n",
        "operationId": "updateTestimonial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestimonialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      },
      "post": {
        "tags": ["create-testimonial-processor"],
        "summary": "Create testimonial",
        "description": "1. Create a testimonial for specific organization.\n2. Request code first.\n3. Only consumer have access to this endpoint.\n",
        "operationId": "createTestimonial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestimonialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/v1/testimonial/my-list": {
      "get": {
        "tags": ["list-testimonial-processor"],
        "summary": "List testimonials of organization",
        "description": "1. List testimonials of current authenticated organization.\n2. Only organization have access to this endpoint.\n",
        "operationId": "listMyTestimonials",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "createdTo",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "minRating",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "maxRating",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "sortType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["NEWEST", "OLDEST", "MAX", "MINE"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/testimonial/list": {
      "get": {
        "tags": ["list-testimonial-processor"],
        "summary": "List testimonials of organization",
        "description": "1. List testimonials of specific organization.\n2. Only active consumer and organization have access to this endpoint.\n",
        "operationId": "listTestimonials",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "createdTo",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "minRating",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "maxRating",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "sortType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["NEWEST", "OLDEST", "MAX", "MINE"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateTestimonialRequest": {
        "type": "object",
        "properties": {
          "testimonialId": { "type": "string", "format": "uuid" },
          "rating": { "type": "integer", "format": "int32" },
          "comment": { "type": "string" },
          "code": { "type": "string" }
        }
      },
...
      "CreateTestimonialRequest": {
        "type": "object",
        "properties": {
          "orgId": { "type": "string", "format": "uuid" },
          "rating": { "type": "integer", "format": "int32" },
          "comment": { "type": "string" },
          "verificationCode": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Paste your IdToken here",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}